Fix encoding issue with unicode-encoded device name
My device is called "Küche". The current implementation could not correctly cope with it, neither locally in my win machine, nor on my docker-based home-assistant installation. The proposed PR fixes that on my local machine, but I'm not sure if this is a correct fix for all environments. So handle with care :)